+++ includes/Supports/FormFields/Textarea.php
@@ -14,6 +14,8 @@ class Textarea extends BaseField {
 	 */
 	public function render(): string {
 		$this->set_setting( 'type', 'textarea' );
+		$input_class = $this->get_setting( 'field_class' );
+		$this->set_setting( 'field_class', str_replace( 'sp-input-text', 'sp-input-textarea', $input_class ) );
 
 		return '<textarea ' . $this->build_attributes() . '>' . esc_textarea( $this->get_value() ) . '</textarea>';
 	}