+++ modules/HeroCarousel/Admin.php
@@ -40,7 +40,7 @@ class Admin {
 	/**
 	 * Load meta box content
 	 *
-	 * @param int    $slider_id The slider id.
+	 * @param int $slider_id The slider id.
 	 * @param string $slide_type The slider type.
 	 */
 	public function meta_box_content( int $slider_id, string $slide_type ) {
@@ -56,22 +56,15 @@ class Admin {
 			<?php
 			$content_sliders  = get_post_meta( $post->ID, '_content_slider', true );
 			$content_settings = get_post_meta( $post->ID, '_content_slider_settings', true );
+			$content_settings = is_array( $content_settings ) ? $content_settings : [];
 
 			if ( is_array( $content_sliders ) && count( $content_sliders ) > 0 ) {
 				$total_sliders = count( $content_sliders );
 				foreach ( $content_sliders as $slide_num => $content_slider ) {
-					$item = new Item(
-						$content_slider,
-						array_merge(
-							$content_settings,
-							[
-								'item_id'         => $slide_num,
-								'slider_id'       => $post->ID,
-								'total_items'     => $total_sliders,
-								'lazy_load_image' => true,
-							]
-						)
-					);
+					$item = new Item( $content_slider, $content_settings );
+					$item->set_setting( new Setting( $post->ID ) );
+					$item->set_prop( 'id', $slide_num + 1 );
+					$item->set_prop( 'total_items', $total_sliders );
 
 					self::item_meta_box( $item, $total_sliders );
 				}
@@ -87,10 +80,10 @@ class Admin {
 	 * Item meta box
 	 *
 	 * @param Item $item The Item object.
-	 * @param int  $total_items Total items.
+	 * @param int $total_items Total items.
 	 */
 	public static function item_meta_box( Item $item, int $total_items = 0 ) {
-		$title       = sprintf( '%s %s', __( 'Slide', 'carousel-slider' ), $item->get_item_id() + 1 );
+		$title       = sprintf( '%s %s', __( 'Slide', 'carousel-slider' ), $item->get_item_id() );
 		$action_html = self::get_actions_html( $item->get_slider_id(), $item->get_item_id(), $total_items );
 		?>
 		<div class="shapla-toggle shapla-toggle--normal" data-id="closed">